Move print_model_sparsity to utils
